requirements.txt
django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.7
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.0
setuptools==75.3.0
django-filter==24.3